Update in CustomWaveViewer.cs (See Below) * Added method to fit wave graph to window. * Added control resize method (lagy and a bit buggy). * Added back in visual representation of the two chew count methods. * readData method and the onpaint method now use different sampling sizes (look in global custom variables to change the sample size for the readData method).
~~~~ TO DO ~~~~
We still don't have a way to proove that the algorithms are behaving correctly (we need to create a timestamp validation method).

The First method appears to not be behaving correctly. this could (and probably is) be due to the lines drawn not matching up to the peaks correctly (re-scaling issue / coord noting issue).